--- src/main/java/com/backend/farmon/reposiotry/ChatRoomReposiotry/ChatRoomRepository.java
@@ -1,12 +1,0 @@
-package com.backend.farmon.reposiotry.ChatRoomReposiotry;
-
-import com.backend.farmon.domain.ChatRoom;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long>, ChatRoomRepositoryCustom {
-
-}
