--- src/main/java/com/backend/farmon/reposiotry/EstimateRepository/EstimateRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.backend.farmon.reposiotry.EstimateRepository;
-
-public interface EstimateRepositoryCustom {
-}
