--- src/main/java/com/backend/farmon/reposiotry/ExpertReposiotry/ExpertRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.backend.farmon.reposiotry.ExpertReposiotry;
-
-public interface ExpertRepositoryCustom {
-}
