--- src/main/java/com/backend/farmon/reposiotry/UserRepository/UserRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.backend.farmon.reposiotry.UserRepository;
-
-public interface UserRepositoryCustom {
-}
