+++ src/main/java/com/backend/farmon/repository/ChatRoomReposiotry/ChatRoomRepository.java
@@ -1,0 +1,8 @@
+package com.backend.farmon.repository.ChatRoomReposiotry;
+
+import com.backend.farmon.domain.ChatRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long>, ChatRoomRepositoryCustom {
+
+}
