+++ src/main/java/com/backend/farmon/repository/ChatRoomReposiotry/ChatRoomRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.backend.farmon.reposiotry.ChatRoomReposiotry;
+package com.backend.farmon.repository.ChatRoomReposiotry;
 
 import com.backend.farmon.domain.ChatRoom;
 import org.springframework.data.domain.Page;
