+++ src/main/java/com/backend/farmon/repository/EstimateRepository/EstimateRepository.java
@@ -1,9 +1,7 @@
-package com.backend.farmon.reposiotry.EstimateRepository;
+package com.backend.farmon.repository.EstimateRepository;
 
 import com.backend.farmon.domain.Estimate;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface EstimateRepository extends JpaRepository<Estimate, Long>, EstimateRepositoryCustom {
 }
