+++ src/main/java/com/backend/farmon/repository/EstimateRepository/EstimateRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.backend.farmon.repository.EstimateRepository;
+
+public interface EstimateRepositoryCustom {
+}
