+++ src/main/java/com/backend/farmon/repository/ExpertReposiotry/ExpertRepository.java
@@ -1,4 +1,4 @@
-package com.backend.farmon.reposiotry.ExpertReposiotry;
+package com.backend.farmon.repository.ExpertReposiotry;
 
 import com.backend.farmon.domain.Expert;
 import org.springframework.data.jpa.repository.JpaRepository;
