+++ src/main/java/com/backend/farmon/repository/ExpertReposiotry/ExpertRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.backend.farmon.repository.ExpertReposiotry;
+
+public interface ExpertRepositoryCustom {
+}
