+++ src/main/java/com/backend/farmon/repository/UserRepository/UserRepository.java
@@ -1,8 +1,6 @@
-package com.backend.farmon.reposiotry.UserRepository;
+package com.backend.farmon.repository.UserRepository;
 
-import com.backend.farmon.domain.Expert;
 import com.backend.farmon.domain.User;
-import com.backend.farmon.reposiotry.ExpertReposiotry.ExpertRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryCustom {
