+++ src/main/java/com/backend/farmon/repository/UserRepository/UserRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.backend.farmon.repository.UserRepository;
+
+public interface UserRepositoryCustom {
+}
