+++ src/main/java/com/backend/farmon/service/ChatRoomService/ChatRoomQueryServiceImpl.java
@@ -5,9 +5,9 @@ import com.backend.farmon.apiPayload.exception.handler.EstimateHandler;
 import com.backend.farmon.converter.ChatConverter;
 import com.backend.farmon.domain.*;
 import com.backend.farmon.dto.chat.ChatResponse;
-import com.backend.farmon.reposiotry.ChatMessageRepository.ChatMessageRepository;
-import com.backend.farmon.reposiotry.ChatRoomReposiotry.ChatRoomRepository;
-import com.backend.farmon.reposiotry.UserRepository.UserRepository;
+import com.backend.farmon.repository.ChatMessageRepository.ChatMessageRepository;
+import com.backend.farmon.repository.ChatRoomReposiotry.ChatRoomRepository;
+import com.backend.farmon.repository.UserRepository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
